Add support for 1-Wire overdrive mode
Some 1-Wire devices, such as the DS28E15 authenticator, do not operate
at the standard speed which the BP currently supports. They require the
faster variant known as "overdrive" in order to even acknowledge their
presence. This changeset adds support for this mode, including the
ability to select it via the binary command mode.