Extract common message strings into a single file.
Keep in mind the #include facility in the message files is extremely
brittle and it only works if the included files are in the same
directory as the Python script.  Since this script is ran only
occasionally, there is no immediate need to increase its robustness.

This also had the benefit that v4 strings will not be included in v3
firmwares anymore, allowing for more code to be included on those tiny
little boards...