Fix printing zero on console (#91)
Not sure if this is the best way to fix it but this worked for me.
If zero is passed to the bp_write_dec_* functions to be printed then nothing will end up on screen.
This can be see if you try to run the self test and no errors are reported.
ex. "Found  errors."
A simple fix was to just check for zero and handle it explicitly.